+++ src/game_file.rs
@@ -8,14 +8,12 @@ impl GameFile {
         main_code: [u8; MAIN_CODE],
         code_banks: Vec<[u8; CODE_BANK]>,
         atlases: Vec<[u8; ATLAS_BANK]>,
-        controller_graphics: Vec<[u8; CONTROLLER_GRAPHICS_BANK]>,
     ) -> Self {
         Self {
             header,
             main_code,
             code_banks,
             atlases,
-            controller_graphics,
         }
     }
 }
@@ -28,14 +26,11 @@ impl Readable for GameFile {
             .map_err(|e| FileAccessError(e, "reading main code"))?;
         let code_banks = read_sized_blocks(reader, header.code_bank_count as usize)?;
         let atlas_banks = read_sized_blocks(reader, header.atlas_bank_count as usize)?;
-        let controller_banks =
-            read_sized_blocks(reader, header.controller_graphics_bank_count as usize)?;
         Ok(GameFile::new(
             header,
             convert_vec(main_code),
             code_banks,
             atlas_banks,
-            controller_banks,
         ))
     }
 }
@@ -48,9 +43,6 @@ impl Writeable for GameFile {
             output.extend_from_slice(bank);
         }
         for bank in &self.atlases {
-            output.extend_from_slice(bank);
-        }
-        for bank in &self.controller_graphics {
             output.extend_from_slice(bank);
         }
 
@@ -77,24 +69,27 @@ fn read_sized_blocks<R: ReaderExt, const N: usize>(
 }
 
 fn convert_vec<T, const N: usize>(v: Vec<T>) -> [T; N] {
-    v.try_into()
-        .unwrap_or_else(|v: Vec<T>| panic!("Expected a Vec of length {} but it was {}", N, v.len()))
+    v.try_into().unwrap_or_else(|v: Vec<T>| {
+        panic!(
+            "Expected a Vec of length {} but it was {} (please create github issue)",
+            N,
+            v.len()
+        )
+    })
 }
 
 #[cfg(test)]
 mod test {
     use crate::read_write_impl::Writeable;
-    use crate::{
-        GameFile, GameFileHeader, ATLAS_BANK, CODE_BANK, CONTROLLER_GRAPHICS_BANK, MAIN_CODE,
-    };
+    use crate::{GameFile, GameFileHeader, ATLAS_BANK, CODE_BANK, MAIN_CODE};
 
     #[test]
     #[rustfmt::skip]
     fn test_write() {
-        let header = GameFileHeader::new(String::from("1"), 1, 1, 1,0, String::from("a"), String::from("b"), String::from("c"), 1, 1, 1);
-        let file = GameFile::new(header, [1; MAIN_CODE], vec![[2; CODE_BANK]], vec![[3; ATLAS_BANK]], vec![[4; CONTROLLER_GRAPHICS_BANK]]);
+        let header = GameFileHeader::new(String::from("1"), 1, 1, 1,0, String::from("a"), String::from("b"), String::from("c"), 1, 1);
+        let file = GameFile::new(header, [1; MAIN_CODE], vec![[2; CODE_BANK]], vec![[3; ATLAS_BANK]]);
         
         let bytes  = file.as_bytes().unwrap();
-        assert_eq!(bytes.len(), MAIN_CODE + CODE_BANK + ATLAS_BANK + CONTROLLER_GRAPHICS_BANK + 23);
+        assert_eq!(bytes.len(), MAIN_CODE + CODE_BANK + ATLAS_BANK + 22);
     }
 }